SLING-8743 Move DistributionQueueBuilderFactory class from impl to api surface
