+++ src/main/java/org/apache/sling/distribution/queue/DistributionQueueProviderFactory.java
@@ -17,12 +17,34 @@
  * under the License.
  */
 
-package org.apache.sling.distribution.queue.impl;
+package org.apache.sling.distribution.queue;
 
+import aQute.bnd.annotation.ProviderType;
+import org.apache.sling.distribution.queue.spi.DistributionQueueProvider;
+
+/**
+ * Allows obtaining service instances of {@link DistributionQueueProviderFactory} as regsitered with
+ * the OSGi runtime
+ */
+
+@ProviderType
 public interface DistributionQueueProviderFactory {
 
+    /**
+     * given the agent and service name, obtain a provider of {@link DistributionQueueProvider}
+     *
+     * @param agentName the name of the agent this queue is intended to be used for
+     * @param serviceName name to the service used to obtain repository sessions of (if needed)
+     * @return a DistributionQueueProvider instance, or {@code null} if none could be obtained
+     */
     DistributionQueueProvider getProvider(String agentName, String serviceName);
 
+    /**
+     * release (and clean up) the backing resources (compute or otherwise) of the
+     * {@link DistributionQueueProvider} supplied
+     *
+     * @param queueProvider the {@link DistributionQueueProvider} to be released
+     */
     void releaseProvider(DistributionQueueProvider queueProvider);
 
 }
