+++ src/main/java/org/apache/sling/distribution/queue/spi/DistributionQueueProvider.java
@@ -16,9 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.sling.distribution.queue.impl;
+package org.apache.sling.distribution.queue.spi;
 
-import aQute.bnd.annotation.ProviderType;
+import aQute.bnd.annotation.ConsumerType;
 import org.apache.sling.distribution.common.DistributionException;
 import org.apache.sling.distribution.queue.DistributionQueueType;
 import org.apache.sling.distribution.queue.spi.DistributionQueue;
@@ -27,7 +27,7 @@ import org.jetbrains.annotations.NotNull;
 /**
  * A provider for {@link DistributionQueue}s
  */
-@ProviderType
+@ConsumerType
 public interface DistributionQueueProvider {
 
     /**
